fixup! Improve readability be extracting separate types to their own files
Move StdioOutputStream to same file as StreamLogHandler, and extract related tests to separate file
